notes are now invisible until past the dancer
NotesControl now prevents the notes from appearing until they've passed
the dancer (plus a threshold for margin of error when they move around)
to prevent culling through the dancer. This is currently finding the
dancer's position by the CharacterModel's name. Optimally, there should
be a unique tag for this. However, the scene will require being altered.